updated cli

+++ build/lib/hemlock_cli/deploy.sh
@@ -3,14 +3,14 @@
 cmd__deploy() {
     # Deploy application
     echo "Deploying algorithm"
-    export `python3 $DIR/env/export_yaml.py env/production-scale.yaml`
+    export `python3 $DIR/env/export_yml.py env/production-env.yml`
+    export `python3 $DIR/env/export_yml.py env/production-scale.yml`
     export_lite_vars
     create_app
     create_addons
-    if [ $USE_BUCKET != 0 ]; then
-        set_bucket_cors
-    fi
+    set_bucket_cors
     push_slug
+    heroku git:remote -a $app
     scale
 }
 
@@ -32,9 +32,7 @@ create_app() {
     echo "Creating application"
     heroku apps:create $app
     URL_ROOT=http://$app.herokuapp.com
-    python3 $DIR/env/update_yaml.py env/production-env.yaml URL_ROOT $URL_ROOT
-    heroku config:set \
-        `python3 $DIR/env/export_yaml.py env/production-env.yaml`
+    python3 $DIR/env/update_yml.py env/production-env.yml URL_ROOT $URL_ROOT
     heroku buildpacks:add heroku/python
     heroku buildpacks:add https://github.com/heroku/heroku-buildpack-chromedriver
     heroku buildpacks:add https://github.com/heroku/heroku-buildpack-google-chrome
@@ -45,17 +43,19 @@ create_addons() {
     echo
     echo "Creating postgres and redis addons"
     heroku addons:add heroku-postgresql:$POSTGRES_PLAN
-    # if [ $WORKER_SCALE != 0 ]; then
-    #     heroku addons:add heroku-redis:$REDIS_PLAN
-    # fi
-    heroku addons:add heroku-redis:$REDIS_PLAN
+    if [ $WORKER_SCALE != 0 ]; then
+        heroku addons:add heroku-redis:$REDIS_PLAN
+    fi
 }
 
 set_bucket_cors() {
     # Set production bucket CORS permissions
     echo
-    echo "Setting CORS permissions for production bucket"
-    echo "Enabling bucket $BUCKET CORS permissions for origin $URL_ROOT"
+    if [ -z "$BUCKET" ]; then
+        echo "No bucket detected"
+        return
+    fi
+    echo "Setting CORS permissions on bucket $BUCKET for origin $URL_ROOT"
     python3 $DIR/gcloud/create_cors.py $URL_ROOT
     gsutil cors set cors.json gs://$BUCKET
     rm cors.json
@@ -65,10 +65,15 @@ push_slug() {
     # Push Heroku slug
     echo
     echo "Pushing Heroku slug"
+    heroku config:set \
+        `python3 $DIR/env/export_yml.py env/production-env.yml`
     git add .
+    if [ ! -z "$BUCKET" ]; then
+        git add -f env/gcp-credentials.json
+    fi
     git commit -m "deploying survey"
-    git push heroku master
-    heroku git:remote -a $app
+    git push -f heroku master
+    git reset HEAD^
 }
 
 scale() {
@@ -87,7 +92,8 @@ cmd__production() {
     echo "WARNING: This action will override the current database"
     echo "Confirm the application name below to proceed"
     echo
-    export `python3 $DIR/env/export_yaml.py env/production-scale.yaml`
+    export `python3 $DIR/env/export_yml.py env/production-scale.yml`
+    heroku config:set NO_DEBUG_FUNCTIONS=1
     heroku addons:destroy heroku-postgresql
     heroku addons:destroy heroku-redis
     create_addons
@@ -97,23 +103,25 @@ cmd__production() {
 cmd__update() {
     # Update application
     echo "Updating application"
-    heroku config:set `python3 $DIR/env/export_yaml.py env/production-env.yaml`
-    git add .
-    git commit -m "update"
-    git push heroku master
+    export `python3 $DIR/env/export_yml.py env/production-env.yml`
+    set_bucket_cors
+    push_slug
+}
+
+cmd__restart() {
+    # Restart application
+    echo "Restarting application"
+    export `python3 $DIR/env/export_yml.py env/production-env.yml`
+    set_bucket_cors
+    heroku restart
 }
 
 cmd__destroy() {
     # Destroy applicaiton
     echo "Preparing to destroy application"
-    export `python3 $DIR/env/export_yaml.py env/production-env.yaml`
-    if [ $USE_BUCKET != 0 ]; then
-        echo
-        echo "Restricting CORS permissions for production bucket"
-        python3 $DIR/gcloud/create_cors.py ""
-        gsutil cors set cors.json gs://$BUCKET
-        rm cors.json
-    fi
+    python3 $DIR/env/update_yml.py env/production-env.yml URL_ROOT ""
+    export `python3 $DIR/env/export_yml.py env/production-env.yml`
+    set_bucket_cors
     echo
     echo "Destroying application"
     heroku apps:destroy
